refactor: move workspaceselectors
