+++ src/web/src/__tests__/integration/WorkspaceSelector.integration.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, vi } from "vitest";
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { render } from "../test-utils";
-import WorkspaceSelector from "../../views/workspace/WorkspaceSelector";
+import WorkspaceSelector from "../../views/workspace/WorkspaceInstruction/WorkspaceSelector";
 
 // Mock window.location for navigation testing
 const mockLocation = {
