+++ src/web/src/views/workspace/WorkspaceInstruction/WorkspaceCreateDialog.tsx
@@ -10,10 +10,10 @@ import {
   Alert,
 } from "@mui/material";
 import React, { useState, useEffect, useCallback } from "react";
-import SwaggerItemSelector from "./common/SwaggerItemSelector";
+import SwaggerItemSelector from "../common/SwaggerItemSelector";
 import styled from "@emotion/styled";
-import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
-import type { Plane } from "./interfaces";
+import { workspaceApi, specsApi, errorHandlerApi } from "../../../services";
+import type { Plane } from "../interfaces";
 
 interface WorkspaceCreateDialogProps {
   openDialog: boolean;
