+++ src/web/src/views/workspace/WorkspaceInstruction/index.ts
@@ -1,0 +1,6 @@
+export { default as WorkspaceInstruction } from "./WorkspaceInstruction";
+export { default as WorkspaceSelector } from "./WorkspaceSelector";
+export { default as WorkspaceCreateDialog } from "./WorkspaceCreateDialog";
+
+import WorkspaceInstruction from "./WorkspaceInstruction";
+export default WorkspaceInstruction;
